requirements.txt
argparse==1.1
logging==0.5.1.2
tensorflow==2.6.2
torch==1.4.0
attrs==20.3.0
json==2.0.9
habitat-sim==0.1.4
habitat==0.1.4
librosa==0.9.1
numpy==1.19.5
networkx==2.2
scipy==1.5.4
gym==0.10.9
cv2==4.5.1
tqdm==4.56.0
pyroomacoustics==0.4.2
SoundFile==0.10.3.post1
Pillow==8.1.0
scikit-image==0.16.2
torchvision==0.5.0
torchsummary==1.5.1
